[Ostro OS XT] Fallback MADV_FREE to MADV_DONTNEED in madvise
It fixes XWALK-7302.

The root cause is when running on Ostro OS XT, the MADV_FREE
(defined in glibc 2.24) is passed to the syscall madvise(),
while the kernel (4.4) cannot understand the value. This
then triggers a RELEASE_ASSERT.

This patch fallbacks MADV_FREE to MADV_DONTNEED before the issue
gets fixed in Ostro OS XT.

BUG=XWALK-7302

TODO=Fix in Ostro OS XT upstream