Revert "[Ostro OS XT] Set CAP_MAC_OVERRIDE capability for renderer process"
This reverts commit cf00507599f4a0b1c4f3facf93c210d9f086f159.

With the kernel bug fixed[1][2], it's no longer necessary to carry this
problematic workaround in meta-crosswalk. Please refer to
https://github.com/ostroproject/ostro-os-xt/issues/43 for more details.

Previously xwalk must be started with the `--no-sandbox` option since
not all threads will be granted with the CAP_MAC_OVERRIDE capability
when sandbox is enabled, even with the reverted commit applied. This
option is not required now after the kernel bug fix.

[1] http://git.yoctoproject.org/cgit/cgit.cgi/linux-yocto-4.4/commit/?h=standard/intel/bxt-rebase&id=661ce9abb6a53b5e07e7b3284ea072099e192293
[2] https://github.com/ostroproject/ostro-os-xt/commit/6d8e8d5e4680ef083cb2029e5ebe13765a57f5a7

BUG=XWALK-7314
BUG=XWALK-7322